[AU4] changed file struct of projectscene module
