+++ au4/src/projectscene/projectscene.qrc
@@ -8,8 +8,8 @@
         <file>qml/Audacity/ProjectScene/internal/audio/KnobControl.qml</file>
         <file>qml/Audacity/ProjectScene/internal/audio/VolumePressureMeter.qml</file>
         <file>qml/Audacity/ProjectScene/internal/TracksControlPanel.qml</file>
-        <file>qml/Audacity/ProjectScene/clipsview_v1/TracksClipsView.qml</file>
-        <file>qml/Audacity/ProjectScene/clipsview_v1/TrackClipsItem.qml</file>
-        <file>qml/Audacity/ProjectScene/clipsview_v1/ClipItem.qml</file>
+        <file>qml/Audacity/ProjectScene/clipsview/TracksClipsView.qml</file>
+        <file>qml/Audacity/ProjectScene/clipsview/TrackClipsItem.qml</file>
+        <file>qml/Audacity/ProjectScene/clipsview/ClipItem.qml</file>
     </qresource>
 </RCC>
