--- au4/src/projectscene/view/clipsview/trackclipsitem.cpp
@@ -1,18 +1,0 @@
-#include "trackclipsitem.h"
-
-using namespace au::projectscene;
-
-TrackClipsItem::TrackClipsItem(QObject* parent)
-    : QObject(parent)
-{}
-
-QList<ClipItem*> TrackClipsItem::clips() const
-{
-    return m_clips;
-}
-
-void TrackClipsItem::setClips(QList<ClipItem*> list)
-{
-    m_clips = list;
-    emit clipsChanged();
-}
