--- au4/src/projectscene/view/clipsview_v1/clipkey.cpp
@@ -1,7 +1,0 @@
-#include "clipkey.h"
-
-using namespace au::projectscene;
-
-ClipKey::ClipKey()
-{
-}
