--- au4/src/projectscene/view/clipsview_v1/clipkey.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include <QObject>
-#include "processing/processingtypes.h"
-
-namespace au::projectscene {
-class ClipKey
-{
-    Q_GADGET
-
-public:
-    ClipKey();
-
-    processing::ClipKey key;
-};
-}
