--- au4/src/projectscene/view/clipsview_v1/clipsview_v1.cmake
@@ -1,15 +1,0 @@
-
-
-set(CLIPSVIEW_V1_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/trackslistclipsmodel.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/trackslistclipsmodel.h
-    ${CMAKE_CURRENT_LIST_DIR}/clipslistmodel.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/clipslistmodel.h
-    ${CMAKE_CURRENT_LIST_DIR}/clipkey.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/clipkey.h
-
-    ${CMAKE_CURRENT_LIST_DIR}/TimelineContext.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/TimelineContext.h
-    ${CMAKE_CURRENT_LIST_DIR}/WaveView.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/WaveView.h
-)
